chore: refactor Vuex printers state to Pinia printers store
