+++ client-vue/src/stores-pinia/outlet-current.store.ts
@@ -1,12 +1,32 @@
 import { defineStore } from "pinia";
 import { OutletCurrentValues } from "@/models/sse-messages/printer-sse-message.model";
 
+interface State {
+  outletCurrentValues?: OutletCurrentValues;
+}
+
 export const useOutletCurrentStore = defineStore("OutletCurrent", {
-  state: () => ({
-    outletCurrentValues: undefined as OutletCurrentValues | undefined,
+  state: (): State => ({
+    outletCurrentValues: undefined,
   }),
+  getters: {
+    someVal: (state) => {
+      return !!state.outletCurrentValues;
+    },
+    someParamVal: (state) => {
+      return (val: boolean) => val || !state.outletCurrentValues;
+    },
+    otherVal() {
+      return this.someVal || this.someParamVal(1);
+    },
+    otherParamVal() {
+      return (val2: boolean) => {
+        this.someParamVal(val2);
+      };
+    },
+  },
   actions: {
-    setOutletCurrentValues(outletCurrentValues: OutletCurrentValues) {
+    setOutletCurrentValues(outletCurrentValues?: OutletCurrentValues) {
       this.outletCurrentValues = outletCurrentValues;
     },
   },
